The uncommenting step 1
11 errors in the tests gg

+++ hexarena/Scripts/hexarena/GameLogic/Abilities/MalevolentShrine.cs
@@ -7,7 +7,7 @@ namespace GameLogic
 	
 	public class MalevolentShrine : Passive //imeto e mnogo edgy
 	{
-		private short damage;
+		private int damage;
 		public MalevolentShrine(EventManager eventManager, AllEnemiesTarget target){
 			damage = 10;
 			Connect(eventManager);
