+++ hexarena/Scripts/hexarena/GameLogic/Abilities/PitchforkPoke.cs
@@ -7,7 +7,7 @@ namespace GameLogic{
 //BasicAttack
 public partial class PitchforkPoke : Active
 {
-	private short damage = 1;
+	private int damage = 1;
 	
 	public PitchforkPoke(EventManager _eventManager, SingleTarget _targeting){ 
 		Connect(_eventManager);
